[#12901] Check if the search service is active in account request search test (#13101)
* added check for active search service in account request search test

* moved account request search test to AccountRequestSearchIT and updated variables and asserts accordingly

* Add caution note to instructor email copies and remove hyperlinks

* updated account request test constructor to be consistent with latest commits

* fixed style errors

* fixed compiler errors

* fixed assertion error and style

* fixed linter errors

* Revert "Add caution note to instructor email copies and remove hyperlinks"

This reverts commit 232c75779c3c8a9f49a8648693210bec2dd28b8d.

---------
